refactor: optimiser sidebar
migrate to tsx
deduplicate display code
