+++ src/lib/constants.ts
@@ -742,6 +742,7 @@ export const COMBAT_STATS = 'Combat Stats'
 export const COMPUTE_ENGINE_CPU = 'CPU'
 export const COMPUTE_ENGINE_GPU_STABLE = 'GPU Stable'
 export const COMPUTE_ENGINE_GPU_EXPERIMENTAL = 'GPU Experimental'
+export type ComputeEngine = typeof COMPUTE_ENGINE_CPU | typeof COMPUTE_ENGINE_GPU_EXPERIMENTAL | typeof COMPUTE_ENGINE_GPU_STABLE
 
 export const SACERDOS_RELIVED_ORDEAL_1_STACK = 'Sacerdos\' Relived Ordeal 1x'
 export const SACERDOS_RELIVED_ORDEAL_2_STACK = 'Sacerdos\' Relived Ordeal 2x'
